feat: 주문시 상품 재고 수정 로직 추가

+++ src/main/java/com/example/flashfrenzy/domain/product/dto/ProductResponseDto.java
@@ -12,7 +12,7 @@ public class ProductResponseDto {
 
     private String image;
 
-    private int price;
+    private Long price;
 
     private String category1; // 대분류  ex) 옷
